requirements.txt
pymongo == 4.2.0
ipykernel
pandas
numpy
matplotlib
seaborn
scikit-learn
pymongo[srv]
xgboost==2.1.0
python-dotenv == 1.0.1
certifi
PyYAML == 6.0.1
dill == 0.3.8
scipy
imblearn
xgboost
-e . #editable mode calls the find_packages()